🚀 [Feat]: Detach the HouseRegisterTemplate and Change the Template1
- 디자인 변경에 맞춰 캐러셀 적용과 템플릿 분리
- HouseRegisterTemplate1 UI 구현
- house database에 floor 추가에 따른 houseFormType과 supabase.ts에 floor 추가
