+++ src/types/house.type.ts
@@ -18,6 +18,7 @@ export const HouseForm = z.object({
     required_error: '공유 주거의 넓이를 평 단위로 입력해주세요.',
   }),
   room_num: z.number(),
+  floor: z.union([z.literal(0), z.literal(1), z.literal(2)]).optional(),
   deposit_price: z.number({
     required_error: '보증금을 만원 단위로 입력해주세요.',
   }),
